+++ lumin/nn/models/model.py
@@ -236,8 +236,12 @@ class Model(AbsModel):
             pred = _get_preds(inputs, callbacks)
         else:
             pred = []
-            for i in range(0, len(inputs)-bs+1, bs): pred.append(_get_preds(inputs[i:i+bs], callbacks))
-            pred.append(_get_preds(inputs[i+bs:], callbacks))
+            if isinstance(inputs, tuple):
+                for i in range(0, len(inputs[1])-bs+1, bs): pred.append(_get_preds((inputs[0][i:i+bs], inputs[1][i:i+bs]), callbacks))
+                pred.append(_get_preds((inputs[0][i+bs:], inputs[1][i+bs:]), callbacks))
+            else:
+                for i in range(0, len(inputs)-bs+1, bs): pred.append(_get_preds(inputs[i:i+bs], callbacks))
+                pred.append(_get_preds(inputs[i+bs:], callbacks))
             pred = np.vstack(pred)
         if as_np:
             if 'multiclass' in self.objective: return np.exp(pred)